KISS interface
  * New configuration panel 'IO'
  * Add power density zones.
  * Query and Set Kiss hardware commands
  * Support bcast rsid over kiss socket
  * limit selectable modem options to 8 bit versions only (kiss i/o mode)
  * Employs simple HDLC encapsulation frame for transmit/receive CRC data
    verification.
  * Uses histogram based threshold detection.
  * Include manual squelch button for KISS op's.
  * User configurable inhibit transmit on busy frequencies for nn seconds
  * Impose transmit wait on valid frame reception
  * Uses dual UDP ports for concurrent operation of programs using the
    same IP address.
  * Clears the transmit buffer when data is available for more then 10
    mins and has not been transmitted.
  * Provides ax25 decoding. Code hijacked from the FLDIGI packet modem co
    - redirect character data display on the RX panel
  * Add non standard KISS frame to handle ARQ data.
    -  interface can handle both ARQ data and HDLC ax25 frames on
       the same port.
  * Squelch related command and control functions to the KISS hardware
    interface.
  * User configuration item to disable CSMA from config panel and kiss
    hardware frame.
  * Additional KISS command to adjust KPSQL detection gain (for testing).
  * Removed unused function in sockets.cxx not compatible with LINUX OS.
    (from John, G8BPQ)
  * New command 'BUSY:<T|F>'
  * Custom bindUDP() for UDP support.
  * Waterfall display gain adjusted based on sample rate changes.
  * Expanded conditions that indicate BUSY:T on command request.
  * Log loopback condition to event log.
  * Restore KPQSL set level position on start up.
  * Conditionally log of the KISS data path strings
  * Added debug filter 'KISS control' to event log.
  * KPSQL threshold button indicator remains ON during transmit inhibit.
    indication of active 'Busy Channel'.
  * Added DEBUG Logging command LOG_HEX(const char *buf, int len);
    D:,... ReadFromHostSocket: ..BUSY:. C0 06 42 55 53 59 3A C0
  * Added thread locks to waterfall.cxx:
    - powerDensity()
    - peakFreq(),
    - powerDensityMaximum()
    These are being accessed by more then one thread.
  * Swapped hash.m4 swapped test case to std::tr1::hash from std::hash
    - Updated pskrep.cxx to handle the change in hash.m4
  * 8PSK2000 removed.
  * Coordination merge with branch kiss21.
  * Code clean up
  * Updated Documentation. Remove unused png's.
  * Request from Jim, W2XO: Add log.set_rst_in and log.set_rst_out xmlrpc
    commands.
  * Compiler warning pacification (global in scope).
  * New XMLRPC commands to query and set active IO port.
  * Documentation update for XMLRPC commands
  * WSAStartup and WSACleanup one for one issue.
  * Update Docs for 3.22 release. New png's added.

+++ src/trx/trx.cxx
@@ -6,7 +6,7 @@
 // Copyright (C) 2007-2010
 //		Stelios Bounanos, M0GLD
 //
-// This file is part of fldigi.  Adapted in part from code contained in gmfsk 
+// This file is part of fldigi.  Adapted in part from code contained in gmfsk
 // source code distribution.
 //
 // Fldigi is free software: you can redistribute it and/or modify
@@ -241,7 +241,7 @@ void trx_trx_receive_loop()
 			while (numread < SCBLOCKSIZE && trx_state == STATE_RX)
 				numread += scard->Read(fbuf + numread, SCBLOCKSIZE - numread);
 			if (numread > SCBLOCKSIZE) {
-LOG_ERROR("numread error %d", numread);
+				LOG_ERROR("numread error %lu", (unsigned long) numread);
 				numread = SCBLOCKSIZE;
 			}
 			if (bHighSpeed) {
@@ -331,7 +331,7 @@ void trx_trx_transmit_loop()
 		    return;
 		}
 
-		if ((active_modem != ssb_modem) && 
+		if ((active_modem != ssb_modem) &&
 			(active_modem != anal_modem) &&
 			!active_modem->XMLRPC_CPS_TEST &&
 			!PERFORM_CPS_TEST ) {
@@ -342,7 +342,7 @@ void trx_trx_transmit_loop()
 
 		if (!progdefaults.DTMFstr.empty()) dtmf->send();
 
-		if ( ReedSolomon->assigned(active_modem->get_mode()) && 
+		if ( ReedSolomon->assigned(active_modem->get_mode()) &&
 			 (progdefaults.TransmitRSid || progStatus.n_rsids != 0)) {
 			if (progStatus.n_rsids < 0) {
 				for (int i = 0; i > progStatus.n_rsids; i--) {
@@ -384,7 +384,7 @@ void trx_trx_transmit_loop()
 		} else
 			trx_state = STATE_RX;
 
-		if (ReedSolomon->assigned(active_modem->get_mode()) && 
+		if (ReedSolomon->assigned(active_modem->get_mode()) &&
 			progdefaults.TransmitRSid &&
 			progdefaults.rsid_post &&
 			progStatus.n_rsids >= 0) ReedSolomon->send(false);
@@ -585,7 +585,7 @@ void trx_reset_loop()
 		abort();
 	}
 
-	trx_state = STATE_RX;	
+	trx_state = STATE_RX;
 }
 
 //=============================================================================
@@ -604,7 +604,7 @@ void trx_start(void)
 		LOG(debug::ERROR_LEVEL, debug::LOG_MODEM, "trx already running!");
 		return;
 	}
-	
+
 	if (scard) delete scard;
 	if (ReedSolomon) delete ReedSolomon;
 	if (dtmf) delete dtmf;
@@ -666,7 +666,7 @@ void trx_start(void)
 		LOG(debug::ERROR_LEVEL, debug::LOG_MODEM, "pthread_create failed");
 		trxrunning = false;
 		exit(1);
-	} 
+	}
 	trxrunning = true;
 }
